#179: Refactor
