+++ lib/tools/qr_code/qr_code_tool.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import '../../i18n/strings.g.dart';
 import '../../tool/feature_tool.dart';
 import 'feature/qr_code_feature.dart';
-import 'qr_code_screen.dart';
+import 'ui/qr_code_screen.dart';
 
 final qrCodeTool = FeatureTool(
   titleBuilder: (context) => Translations.of(context).qrCode.title,
